Move array_to_uint64() to dataConversion files
