+++ NG474-BOOT/Core/Inc/canDriver.h
@@ -68,11 +68,4 @@ void can_data_handler(uint32_t Identifier, uint8_t *rxdata_pt);
   */
 void can_host_handler(void);
 
-/**
-  * @brief	Converts an array of 8 uint8_t elements to a uint64_t
-  * @param	Array is an array of 8 uint8_t elements
-  * @retval	uint64_t converted value
-  */
-uint64_t array_to_uint64(uint8_t *Array);
-
 #endif /* INC_CANDRIVER_H_ */
