+++ NG474-BOOT/Core/Inc/dataConversion.h
@@ -13,6 +13,11 @@
 #define INC_DATACONVERSION_H_
 
 /* Private function prototypes -----------------------------------------------*/
-
+/**
+  * @brief	Converts an array of 8 uint8_t elements to a uint64_t
+  * @param	Array is an array of 8 uint8_t elements
+  * @retval	uint64_t converted value
+  */
+uint64_t array_to_uint64(uint8_t *Array);
 
 #endif /* INC_DATACONVERSION_H_ */
