+++ NG474-BOOT/Core/Src/canDriver.c
@@ -13,6 +13,7 @@
 #include "canDriver.h"
 #include "bootloader.h"
 #include "canReplyMsg.h"
+#include "dataConversion.h"
 
 /* Variables declared elsewhere-----------------------------------------------*/
 extern FDCAN_HandleTypeDef hfdcan1; // declared in fdcan.c
@@ -132,24 +133,6 @@ void can_filter_init(void)
 	{
 		Error_Handler();
 	}
-}
-
-/**
-  * @brief	Converts an array of 8 uint8_t elements to a uint64_t
-  * @param	Array is an array of 8 uint8_t elements
-  * @retval	uint64_t converted value
-  */
-uint64_t array_to_uint64(uint8_t *Array)
-{
-	int array_index;
-	uint64_t converted_value = 0;
-	for (array_index = 0; array_index < 7; array_index++)
-	{
-	  converted_value += Array[array_index];
-	  converted_value = (converted_value << 8);
-	}
-	converted_value += Array[7];
-	return converted_value;
 }
 
 #if USE_TX_INTERRUPT > 0
